# Batch market data requests to IB