--- a/ibkr.py
+++ b/ibkr.py
@@ -6,7 +6,7 @@ from model import Currency, Cash, Instrument, Stock, Bond, Option, OptionType, F
 from parsetools import lenientParse
 from pathlib import Path
 from progress.spinner import Spinner
-from typing import Any, Awaitable, Callable, Dict, List, NamedTuple, Optional, Type, no_type_check
+from typing import Any, Awaitable, Callable, Dict, Iterable, List, NamedTuple, Optional, Tuple, Type, no_type_check
 
 import backoff
 import ib_insync as IB
@@ -528,37 +528,45 @@ class IBDataProvider(MarketDataProvider):
         self._client = client
         super().__init__()
 
-    def fetchQuote(self,
-                   instrument: Instrument,
-                   dataType: MarketDataType = MarketDataType.DELAYED_FROZEN
-                   ) -> Quote:
+    def fetchQuotes(self,
+                    instruments: Iterable[Instrument],
+                    dataType: MarketDataType = MarketDataType.DELAYED_FROZEN
+                    ) -> Iterable[Tuple[Instrument, Quote]]:
         self._client.reqMarketDataType(dataType.value)
 
-        con = contract(instrument)
-        self._client.qualifyContracts(con)
+        # IB.Contract is not guaranteed to be hashable, so we orient the table this way, albeit less useful.
+        # TODO: Check uniqueness of instruments
+        contractsByInstrument: Dict[Instrument, IB.Contract] = {
+            i: contract(i)
+            for i in instruments
+        }
 
-        ticker = self._client.reqTickers(con)[0]
-        logging.info(f'Received ticker: {ticker!r}')
+        self._client.qualifyContracts(*contractsByInstrument.values())
+        tickers = self._client.reqTickers(*contractsByInstrument.values())
 
-        bid: Optional[Cash] = None
-        ask: Optional[Cash] = None
-        last: Optional[Cash] = None
-        close: Optional[Cash] = None
+        for ticker in tickers:
+            instrument = next((i for (i, c) in contractsByInstrument.items()
+                               if c == ticker.contract))
 
-        if (ticker.bid
-                and math.isfinite(ticker.bid)) and not ticker.bidSize == 0:
-            bid = Cash(currency=instrument.currency,
-                       quantity=Decimal(ticker.bid))
-        if (ticker.ask
-                and math.isfinite(ticker.ask)) and not ticker.askSize == 0:
-            ask = Cash(currency=instrument.currency,
-                       quantity=Decimal(ticker.ask))
-        if (ticker.last
-                and math.isfinite(ticker.last)) and not ticker.lastSize == 0:
-            last = Cash(currency=instrument.currency,
-                        quantity=Decimal(ticker.last))
-        if ticker.close and math.isfinite(ticker.close):
-            close = Cash(currency=instrument.currency,
-                         quantity=Decimal(ticker.close))
+            bid: Optional[Cash] = None
+            ask: Optional[Cash] = None
+            last: Optional[Cash] = None
+            close: Optional[Cash] = None
 
-        return Quote(bid=bid, ask=ask, last=last, close=close)
+            if (ticker.bid
+                    and math.isfinite(ticker.bid)) and not ticker.bidSize == 0:
+                bid = Cash(currency=instrument.currency,
+                           quantity=Decimal(ticker.bid))
+            if (ticker.ask
+                    and math.isfinite(ticker.ask)) and not ticker.askSize == 0:
+                ask = Cash(currency=instrument.currency,
+                           quantity=Decimal(ticker.ask))
+            if (ticker.last and math.isfinite(
+                    ticker.last)) and not ticker.lastSize == 0:
+                last = Cash(currency=instrument.currency,
+                            quantity=Decimal(ticker.last))
+            if ticker.close and math.isfinite(ticker.close):
+                close = Cash(currency=instrument.currency,
+                             quantity=Decimal(ticker.close))
+
+            yield (instrument, Quote(bid=bid, ask=ask, last=last, close=close))