--- a/model/__init__.py
+++ b/model/__init__.py
@@ -1,5 +1,5 @@
 from .activity import Activity, DividendPayment, Trade, TradeFlags
 from .cash import Currency, Cash
 from .instrument import Instrument, Stock, Bond, Option, OptionType, FutureOption, Future, Forex
-from .marketdata import Quote, LiveDataProvider
+from .marketdata import Quote, MarketDataProvider
 from .position import Position