Resolves #23. It's a little incredible just how much faster this is. 🚀 

Before (fetching live values for 25 portfolio positions):
```
0.68s user 0.13s system 0% cpu 2:13.61 total
```

After:
```
0.65s user 0.12s system 6% cpu 12.785 total
```